Extracted table into sub component
