+++ package.json
@@ -22,7 +22,7 @@
     "@nivo/core": "^0.85.1",
     "@nivo/pie": "^0.85.1",
     "@tanstack/react-table": "^8.17.3",
-    "@us-gov-cdc/cdc-react": "^1.11.0",
+    "@us-gov-cdc/cdc-react": "^1.12.0",
     "@us-gov-cdc/cdc-react-icons": "^2.5.0",
     "date-fns": "^3.6.0",
     "msw": "^2.2.1",
